ver2

+++ lab11_1.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 //Write prototype of function fibonacci() here
-int fibonacci(int);
+unsigned long long fibonacci(int);
 
 int main(){
 	cout << fibonacci(0) << "\n";
@@ -19,7 +19,7 @@ int main(){
 
 //Write definition of function fibonacci() here
 
-int fibonacci(int x){
+unsigned long long fibonacci(int x){
   if(x==0) return 0;
   else if(x==1) return 1;
   else if(x>1){
